% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_subjects.R
\name{get_all_subjects}
\alias{get_all_subjects}
\title{Get all main subjects of BDL API}
\usage{
get_all_subjects(rate = 60/60, verbose = TRUE)
}
\arguments{
\item{rate}{maximum number of requests per second 60/60 by default}

\item{verbose}{if true then print some info}
}
\value{
a tibble with main subjects ids and their lower levels
}
\description{
Get all main subjects of BDL API
}
\examples{
get_main_subjects(verbose = TRUE)
}
