% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data by variable for a NUTS level}
\usage{
get_data(variable_id, unit_level, rate = 60/60)
}
\arguments{
\item{variable_id}{an id of variable, you can list variables for each
subject using function \code{get_subject_variables()}}

\item{unit_level}{a NUTS level (0 to 6) a numeric value}

\item{rate}{maximum number of requests per second 60/60 by default}
}
\value{
a tibble data and their NUTS ids and names (nuts_id, nuts_name)
}
\description{
Get data by variable for a NUTS level
}
\examples{
get_data("1609969", 3)
}
