% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_main_subjects.R
\name{get_main_subjects}
\alias{get_main_subjects}
\title{Get all main subjects of BDL API}
\usage{
get_main_subjects()
}
\value{
a list with subjects names, ids (tibble) and other subcategories
}
\description{
Get all main subjects of BDL API
}
\examples{
get_main_subjects()
}
