% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts_by_level.R
\name{get_nuts_by_level}
\alias{get_nuts_by_level}
\title{Get Polish NUTS classification}
\usage{
get_nuts_by_level(level, rate = 60/60)
}
\arguments{
\item{level}{(0 to 6) a numeric value}

\item{rate}{maximum number of requests per second 60/60 by default}
}
\value{
a tibble with NUTS classification
}
\description{
Get Polish NUTS classification
}
\examples{
get_nuts_by_level(level = 3)
}
