% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subject_subcategories.R
\name{get_subject_subcategories}
\alias{get_subject_subcategories}
\title{Get all subcategories of main subject}
\usage{
get_subject_subcategories(main_subject_id, sleep_time = 1)
}
\arguments{
\item{main_subject_id}{main subject id e.g K8}

\item{sleep_time}{time in seconds between consecutive queries}
}
\value{
a tibble with all subcategories and variables
}
\description{
Get all subcategories of main subject
}
\examples{
get_main_subjects("K8")
}
