% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subject_variables.R
\name{get_subject_variables}
\alias{get_subject_variables}
\title{Get all subject's variables}
\usage{
get_subject_variables(subject_id, rate = 60/60)
}
\arguments{
\item{rate}{maximum number of requests per second 60/60 by default}

\item{an}{id of a subjects that has variables}
}
\value{
a tibble with variables and their metadata
}
\description{
Get all subject's variables
}
\examples{
get_subject_variables("P1754")
}
