% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_api_token}
\alias{set_api_token}
\title{Set environment variable with BDL API token}
\usage{
set_api_token(token = NULL)
}
\arguments{
\item{token}{a string with BDL API personal token}
}
\value{
nothing
}
\description{
Set environment variable with BDL API token
}
